refactor(solver): split call contextual helpers
